DEV 1.0.1
Fixed issue with SeatPicker where it would not revert back to the green version for the seat type of the seat. Example of this was when a user selected and then deselected a disabled seat, it would show the seat as a standard seat instead of re-showing the green disabled seat.