DEV 1.3
New Booking Wizard
I have built a script that allows us to run a multi-step wizard easily. We simply provide a step name, onload function, onback function and onnext function for each step in the process.

The onLoad function is called when the step is loaded. the onback and OnNext functions are called when the user either navigates forward or back allowing us to intercept this and ensure we can run code to save or revert changes made on the current step.

If the user tries to navigate forward and we don't want them to due to not providing enough or valid information, we can return false in our onNext function which will stop the navigaton continuing and we can handle the error or alert asking for more information.

I will now be starting work on making use of this tool with our main booking page.