DEV 1.5
SOCIAL DISTANCING
We have implemented an algorithm that can determine the seats that need to be blocked off around a users seat selection. This ensures that the cinema can safely offer online booking while maintaining social distancing. The number of seats either side of the users selection that need to be blocked off are configurable.

To accomodate the new social distancing, we have added a new social_distancing column to the gfc_bookings & gfc_films_showtimes tables. the column on the bookings will store the array of seats that should be blocked off. This is stored against the booking so the blocking disappears if the booking is cancelled. The column in the showtimes determines whether the show must have social distancing. This ensures that a show that was given social distancing can retain this throughout its life.

There is a master control for the social distancing as well. When this is turned on, any new showing that is added will have social distancing enabled on it. From the users side, it shows a new seat icon if the seat is blocked and will also state a 'due to social distancing' message if the user clicks on the seat.

This update will help cinemas as they plan to reopen over the next few months.