DEV 3.2
Added the ability to prevent film sales until a particular date. If you set it for a date in advance then the system will display a countdown on the film detail page to countdown to that date. Ticket sales will automatically begin once the sale_unlock date has been reached.