DEV 3.3
Started work on a customer frontend. Users can now login and see their previous bookings which are connected to their email. Users can also reset their password if they forget the password for their account. The reset password system works for both the customer front-end and the Management backend.

We will continue to expand on this to make it more useful.