DEV 5.1.5
Added a check to ensure that the button for the user to view the film detail's text matches the status of the film. If the film is not yet on sale then show as read more.

Added the checks for the online_sales - this ensures we can choose whether online sales of a film or showing can be controlled.