DEV 5.2
Currently working on a new editor.js add-on which will link into the part where we can allow the admin to create pages of content for the front-end website. This is in the early stages and currently just has text and header options. This will be expanded on to allow grid layouts to be created within the